requirements.txt
python-telegram-bot==21.6
yt_dlp
tqdm
requests_toolbelt
psycopg2-binary
